ui: move URL helper into core, trim its tests
